Makefile changes
Separated the third-party library (markdown-it) from the SOURCE files
into its own VENDOR files list.

This may not be the right permanent solution: The vendor files are
concatenated onto the end of the ermrest.js and ermrest.min.js bundles.

But at least this separates them out so that we do not lint or compile
or generate jsdocs from them.