unify template rendering function and its usage
We had a renderTemplate function that was being used directly in
different places. This function would render mustache template with
given data. This data was not following the rules that we put in place
for raw and formatted values (by using '_').
Now renderTemplate function will generate the formatted and raw values
if the given data is not formatted.
Also validateTemplate function has been refactored to be support both
the url_pattern and other usages of mustache template.