Refactor Parser to add ::search:: pseudo-operator
Previously we were using actual ::ciregex:: operator for search, the
problem with that usage was that we were manipulating the searchTerm
internally so ermrest understands it. For instance when we search for
`11`, what we need to send to ermrest is `^(.*[^0-9.])?0*11([^0-9].*|$)`
regular expression. So we needed to keep two versions of filter, one
that users see and one that ermrest understand.

The uri that parser returns (Location.uri) is what the client
(chaise) understand and other apis for ermrest uri (Location.ermrsetUri,
Location.ermrestPath, etc.) are introduced.