Implement FacetColumn, FacetFilters and FacetFilter
FacetColumn will be the objects that client (chaise) should use to
display the facets. FacetFilters is a container for list of
FacetFilters. It will provide client with function for add/remove
different type of filters. It also has a toJSON function which should be
used internally to apply the filters.