Handle Alternative tables in Faceting
Because of alternative logic, the table that detailed is referring to
might be different than compact context. Since we're using the detailed
just for its related entities api, if detailed is actually an alternative
table, it won't have any related entities. Therefore we don't need to handle
that case.
The only case that we need to cover is the case that detailed
is the main table, and compact alternative. In this case, we have to add
linkage from main to alternative to all the detailed related entites while
we're creating the facet columns.