Add alternative support fro faceting
Since we're contextualizing sourceReference in case of entity picker,
we need to have a logic of checking for alternative table before creating
the list of facets. This PR will add that logic. ERMrestjs will check for the
alternative table and if it cannot handle that case, it will just ignore the
facet from facet list.