change _replaceDotWithUnderscore implementations
The assumption for this function was that given object is simple and doesn't
contain circular references. But this is not the case in assets. In this case in client (chaise)
we're attaching extra custom defined objects to the value and this function would get to
an infinite loop. I changed the implementation to:
1. Ignore any custom defined objects.
2. Previously I was mutating the argument, I changed the impelmentation to create a new object
   instead of mutating the old object.

Related Issue: #563