remove sort and paging from csv download link
The current syntax used for sort is something that only read is handling.
Some of the columns provided in the sort criteria might not be actual database
columns, or when you sort based on a column based on heuristics we actually
should sort based on another column. Therefore unless we use the same logic
in Location.ermrestUri we cannot use this API.
One of the challenge of this task is sorting based on pseudo-columns
which we are supporting by adding extra joins,
but in the ermrest entity API this is not easily possible.