add extraAnnotationNames and fix callback usage
This commit will fix the usage of @callback in the documentation. But since
the gccjs doesn't recognize this tag, we have to add it as an extra annotation
that the gccjs should ignore. The original name of this tag was `extra_annotation_name`
which to pass as an argument we should use `extraAnnotationNames`.
The value of this attribute must be an array, otherwise gccjs is ignoring it.
Since the package is using minimist, just passing one value won't turn it into an array.
I didn't know how exactly minimist behaves, but it seems like passing two values does
the work. Therefore I added another tag to the list.