fix FacetColumn.toJSON bug regarding JSON column type
We should make sure the JSON value that we are passingis valid,
if we cannot parse it therefore it should be treated as string literal.
For example if the term is just "test" then the JSON.parse will throw
an error and stringifying it will turn it into "\"test\"" which is
the valid JSON value.