remove reference to facetColumns in refereneCopy
To create new Reference we're using this function in most places.
This function will only do a shallow copy of current Reference, therefore
all the references to the attributes of this Reference object will still remain.
This was causing problem in sort and faceting. Since each facet keeps track of its
Reference, after doing sort (therefore creating a new Reference), the facetColumn.Reference
still refers to the old object. To fix this issue for now, I am deleting the
_facetColumns, to force the new object to create its own facetColumns list.
Technically _referenceColumns should be removed too. Each referenceColumn is keeping track of
its Refernece too. But doing so, caused client to throw "Maximum call stack size exceeded".
This is happening because in ERMrestJS we're blindly using the _referenceCopy while we could
just create a new reference from scratch.