fix source-definitions.fkeys=[] bug in read
the readPath check to use attributegroup was wrong.
if the table has outbound-foreignkeys but the defined source-definitions.fkeys is empty,
we should not use attributegroup and should fall back to entity.
But the code was using attributegroup if table has outbound fks, regardless source-definitions.

This would cause the last part of the code to throw terminal error since it was assuming
that there's an aggregate list of foreignkeys.