change citation implementation + test cases
I realized with the current implementation, we're not handling
all-outbound paths in citation. So I had to move it to the reference
object. This is because the reference.read will call the reference.activeList
prior to having a tuple object and therefore couldn't handle the waitfor
of Citations since we had them on the tuple object.