flatten templateVariables (instead of options.templateVariables)
While doing this, I realize that the change that I made in PseudoColumn
formatPresentation regarding the value that is should return when
PseudoColumn is scalar all-outbound was not correct.
If there's a sourceMarkdownPattern, we should use the passed templateVariables
otherwise we should not pass it and rely on the Column.formatPresentation
to generate the more proper object.