modify Array.clear polyfill to use defineProperty
This is the safer way of defining a property (as oppose to changing
the prototype directly).
The old polyfill was also causing issues with a legacy code that is
using unprotected for .. in loops. Since we were attaching the
clear directly to property, "clear" was one of the possible values
that for .. in loop was returning.