remvoe the check for already injected ermrestjs files
it was causing some issues in other deployments.
in one of the deployments that navbar app is used, it also was including
ermrestjs and because of this logic the second copy of ermrestjs didn't
have the handlebars and other dependencies. This will be resolved with
another PR, for now, I just removed it to avoid terminal errors.