fix versions of npm deps. + remove npm-shrinkwrap
it was only used to fix the mime package version which can be done
by adding it to package.json as well.
I removed it because it was causing issues with different versions
of node. If we want to have such a file we have to keep updating it
with all the packages that are installed but that's not the intention
at all.