make sure vendor files are properly loaded
The querySelector check only tests for existence Of the tag and 
not necessary load status.
So I added a global variable to check whether the File is loaded or not.
If two instances of ermrestjs are loaded in the page, 
there can be a moment where the vendor file is attached to the DOM 
dynamically but not properly loaded yet. And at that moment the second
ermrestjs instance might check for existence of vendor file.