modify default page.content value to check for links
we manually show rowname as link in these places:
- page.content default or when we’re getting the default display of inline tables
(source.markdown_pattern and row_markdown_pattern/page_markdown_pattern
is not defined for the context)
- when we’re getting the default display for foreign key pseudo-columns.
- when we’re generating the list of entity aggregate values.

only in the first case we were not checking for the links, so I changed
the first one to check for the links as well.

we hold off this change because the method that we're using for
checking the link tag is not robust. we're just checking for a regular expression
which might not be accurate. But the inconsistency is the bigger issue that we solved with this commit.