make sure error status messages are always defined
error status message are used as title in chaise. If they are undefined,
chaise will display the default "Terminal Error" case.

ERMrestJS was assuming that this value is always non-empty since it's
coming from XHR response (response.statusText). But it seems like
this variable will not be populated with HTTP/2. So I had to make sure
the statusText has a default.