make sure wait_for processing is keeping the sourcekey
while processing the wait_for list, the code was replacing the given
sourcekey with the definition and therefore the mapping from sourcekey
to the definition was lost. Also the code was passign the source
definition without properly calling clone.

this was causing issues in the path prefix logic where an all-outbound
in the wait_for wasn't properly sharing path with the facets.