make sure fast_filter_source is honored for urls with join
Let's assume we're looking at table "main" that has foreigns key to
tables "F1" and "F2". And both have "fast_filter_source". In this scenario,
user have selected choices in both facets.

The following is the general syntax of getting values for main:

M:=S:main/
<parsed choices of F1>/$M/
<parsed choices of F2>/$M/

and the following is for getting available choices of F1:

T:=S:main/
<parsed choices of F2>/$T
M:=<path from main to F1>/

While for the "path from main to F1" we cannot use the fast_filter_source,
we should honor the fast_filter_source for the "parsed choices of F2"
in both scenarios.

The changes in this commit will make sure that is the case.