make the code and tests more robust regarding errors
the latest postgres update has changed some of the error messages.

none of the changes are breaking the logic that we have for repaiting
the errors. only some expected behavior has changed. also the status
in some conflict errors had a different capitalizations. so instead
of relying on server, I'm now just using the value that we expect
in ermrestjs/chaise.