discard empty search terms
ermrest doesn't support empty string term in `ciregexp` and throws
400 error. also searching for a empty string in `cirexp` doesn't
make sense in the first place.

the logic to "autocomplete" the odd number of quotes was also producing
empty strings. a search term like `term"` was  turned into `ciregexp(,term)`
due to the additional quote that we are adding to it
(`term"` -> `term""` -> `ciregexp(,term)`)