Fix issues in the json facet support (#1025)
When I initially implemented facets, I assumed we would pass the raw JSON objects to the `addChoiceFilters` API. But in some cases, the Chaise doesn't have access to the raw object and passes the string representation instead.

To ensure that both Chaise and ermrestjs consistently handle JSON facets, I ensured that we could pass string representation of facets in this PR. As part of this I also ensured that we're not applying `stringified` on an already JSON-stringified object. Because of this, refreshing the page in Chaise would keep changing the given filter.

I also fixed the `Tuple.uniqueID` API to handle JSON columns properly. This API is supposed to return a string, but we were concatenating the raw values to generate it. In the case of `json`/`jsonb` types, it would return `[object Object]`, which is obviously not what we want.